+++ Favor/Favor/Configuration/Base/BaseTargetType.swift
