+++ Favor/Favor/Configuration/Base/BaseView.swift
