+++ Favor/Favor/Sources/Scenes/Auth/ViewControllers/Password/FindPasswordVC.swift
@@ -33,8 +33,8 @@ final class FindPasswordViewController: BaseViewController, View {
     return textField
   }()
 
-  private lazy var nextButton: LargeFavorButton = {
-    let button = LargeFavorButton(with: .main("다음"))
+  private lazy var nextButton: FavorLargeButton = {
+    let button = FavorLargeButton(with: .main("다음"))
     return button
   }()
