+++ Favor/Favor/Sources/Scenes/Auth/ViewControllers/Password/ValidateEmailCodeVC.swift
@@ -31,8 +31,8 @@ final class ValidateEmailCodeViewController: BaseViewController, View {
     return textField
   }()
 
-  private lazy var nextButton: LargeFavorButton = {
-    let button = LargeFavorButton(with: .main("다음"))
+  private lazy var nextButton: FavorLargeButton = {
+    let button = FavorLargeButton(with: .main("다음"))
     button.isEnabled = false
     return button
   }()
