+++ Favor/Favor/Sources/Scenes/Auth/ViewControllers/SignInVC.swift
@@ -47,9 +47,9 @@ final class SignInViewController: BaseViewController, View {
     return stackView
   }()
   
-  private lazy var signInButton = LargeFavorButton(with: .main("로그인"))
+  private lazy var signInButton = FavorLargeButton(with: .main("로그인"))
   
-  private lazy var findPasswordButton = PlainFavorButton(with: .logIn("비밀번호 찾기"))
+  private lazy var findPasswordButton = FavorPlainButton(with: .logIn("비밀번호 찾기"))
 
   private lazy var buttonStack: UIStackView = {
     let stackView = UIStackView()
