+++ Favor/Favor/Sources/Scenes/Auth/ViewControllers/SignUpVC.swift
@@ -77,8 +77,8 @@ final class SignUpViewController: BaseViewController, View {
     return textField
   }()
   
-  private lazy var nextButton: LargeFavorButton = {
-    let button = LargeFavorButton(with: .main("다음"))
+  private lazy var nextButton: FavorLargeButton = {
+    let button = FavorLargeButton(with: .main("다음"))
     button.configurationUpdateHandler = { button in
       switch button.state {
       case .normal:
