+++ Favor/Favor/Sources/Scenes/Auth/Views/TermCell.swift
@@ -1,21 +1,25 @@
 //
-//  AcceptAllView.swift
+//  TermCell.swift
 //  Favor
 //
-//  Created by 이창준 on 2023/03/03.
+//  Created by 이창준 on 2023/03/02.
 //
 
 import UIKit
 
+import Reusable
 import RxCocoa
 import RxSwift
 import SnapKit
 
-final class AcceptAllView: UIView {
+final class TermCell: UITableViewCell, Reusable {
 
   // MARK: - Constants
 
   // MARK: - Properties
+
+  var disposeBag = DisposeBag()
+  var url: String?
 
   // MARK: - UI Components
 
@@ -33,47 +37,67 @@ final class AcceptAllView: UIView {
   private lazy var titleLabel: UILabel = {
     let label = UILabel()
     label.font = .favorFont(.regular, size: 14)
-    label.text = "약관 전체동의"
     return label
   }()
 
-  fileprivate lazy var button = UIButton()
+  fileprivate lazy var openDetailButton: FavorPlainButton = {
+    let button = FavorPlainButton(with: .more("보기"))
+    return button
+  }()
 
   // MARK: - Initializer
 
-  override init(frame: CGRect) {
-    super.init(frame: frame)
+  override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+    super.init(style: style, reuseIdentifier: reuseIdentifier)
     self.setupStyles()
     self.setupLayouts()
     self.setupConstraints()
+
+    self.openDetailButton.rx.tap
+      .asDriver(onErrorRecover: {_ in return .never()})
+      .drive(with: self, onNext: { owner, tap in
+        guard
+          let urlString = owner.url,
+          let url = URL(string: urlString)
+        else { return }
+
+        UIApplication.shared.open(url)
+      })
+      .disposed(by: self.disposeBag)
   }
 
   required init?(coder: NSCoder) {
     fatalError("init(coder:) has not been implemented")
   }
 
+  // MARK: - Binding
+
+  public func bind(terms: Terms) {
+    self.titleLabel.text = terms.title
+    let image = terms.isAccepted ? "checkmark.square.fill" : "checkmark.square"
+    self.checkButton.configuration?.image = UIImage(systemName: image)
+    self.url = terms.url
+  }
+
   // MARK: - Functions
 
-  public func updateCheckButton(isAllAccepted: Bool) {
-    let image = isAllAccepted ? "checkmark.square.fill" : "checkmark.square"
-    self.checkButton.configuration?.image = UIImage(systemName: image)
-  }
 }
 
 // MARK: - UI Setup
 
-extension AcceptAllView: BaseView {
+extension TermCell: BaseView {
   func setupStyles() {
-    self.frame = self.frame.inset(
+    self.contentView.frame = self.contentView.frame.inset(
       by: UIEdgeInsets(top: 8, left: 0, bottom: 8, right: 0)
     )
+    self.selectionStyle = .none
   }
 
   func setupLayouts() {
     [
       self.checkButton,
       self.titleLabel,
-      self.button
+      self.openDetailButton
     ].forEach {
       self.addSubview($0)
     }
@@ -88,18 +112,13 @@ extension AcceptAllView: BaseView {
     self.titleLabel.snp.makeConstraints { make in
       make.leading.equalTo(self.checkButton.snp.trailing).offset(8)
       make.centerY.equalToSuperview()
-      make.trailing.equalToSuperview()
+      make.trailing.equalTo(self.openDetailButton.snp.leading)
     }
 
-    self.button.snp.makeConstraints { make in
-      make.edges.equalToSuperview()
+    self.openDetailButton.setContentHuggingPriority(.defaultHigh, for: .horizontal)
+    self.openDetailButton.snp.makeConstraints { make in
+      make.trailing.directionalVerticalEdges.equalToSuperview()
+      make.centerY.equalToSuperview()
     }
-  }
-}
-
-extension Reactive where Base: AcceptAllView {
-  var tap: ControlEvent<()> {
-    let source = base.button.rx.tap
-    return ControlEvent(events: source)
   }
 }
