+++ Favor/Favor/Sources/Scenes/Auth/Views/TermTableView.swift
@@ -1,5 +1,5 @@
 //
-//  SelfSizingTableView.swift
+//  TermTableView.swift
 //  Favor
 //
 //  Created by 이창준 on 2023/03/02.
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class SelfSizingTableView: UITableView {
+class TermTableView: UITableView {
   override var contentSize: CGSize {
     didSet {
       self.invalidateIntrinsicContentSize()
