+++ Favor/Favor/Sources/Scenes/Common/Reactors/CellReactors/FavorGiftStatsCellReactor.swift
@@ -1,5 +1,5 @@
 //
-//  GiftStatsCellReactor.swift
+//  FavorGiftStatsCellReactor.swift
 //  Favor
 //
 //  Created by 이창준 on 2023/02/15.
@@ -7,7 +7,7 @@
 
 import ReactorKit
 
-final class GiftStatsCellReactor: Reactor {
+final class FavorGiftStatsCellReactor: Reactor {
   
   // MARK: - Properties
