+++ Favor/Favor/Sources/Scenes/Common/Views/Cells/FavorSetupProfileCell.swift
@@ -45,7 +45,7 @@ class FavorSetupProfileCell: UICollectionViewCell, Reusable, View {
     return label
   }()
   
-  private lazy var goButton = SmallFavorButton(with: .gray("바로가기"))
+  private lazy var goButton = FavorSmallButton(with: .gray("바로가기"))
   
   // MARK: - Initializer
