+++ Favor/Favor/Sources/Scenes/Common/Views/FavorCategoryView.swift
@@ -1,5 +1,5 @@
 //
-//  CategoryView.swift
+//  FavorCategoryView.swift
 //  Favor
 //
 //  Created by 김응철 on 2023/02/06.
@@ -9,17 +9,17 @@ import UIKit
 
 import SnapKit
 
-final class CategoryView: UIScrollView {
+final class FavorCategoryView: UIScrollView {
   
   // MARK: - Properties
   
-  private let lightGiftButton = SmallFavorButton(with: .mainWithIcon("가벼운 선물", imageName: ""))
-  private let birthButton = SmallFavorButton(with: .mainWithIcon("생일", imageName: ""))
-  private let houseWarmingButton = SmallFavorButton(with: .mainWithIcon("집들이", imageName: ""))
-  private let testButton = SmallFavorButton(with: .mainWithIcon("시험", imageName: ""))
-  private let promotionButton = SmallFavorButton(with: .mainWithIcon("승진", imageName: ""))
-  private let graduationButton = SmallFavorButton(with: .mainWithIcon("졸업", imageName: ""))
-  private let etcButton = SmallFavorButton(with: .mainWithIcon("기타", imageName: ""))
+  private let lightGiftButton = FavorSmallButton(with: .mainWithIcon("가벼운 선물", imageName: ""))
+  private let birthButton = FavorSmallButton(with: .mainWithIcon("생일", imageName: ""))
+  private let houseWarmingButton = FavorSmallButton(with: .mainWithIcon("집들이", imageName: ""))
+  private let testButton = FavorSmallButton(with: .mainWithIcon("시험", imageName: ""))
+  private let promotionButton = FavorSmallButton(with: .mainWithIcon("승진", imageName: ""))
+  private let graduationButton = FavorSmallButton(with: .mainWithIcon("졸업", imageName: ""))
+  private let etcButton = FavorSmallButton(with: .mainWithIcon("기타", imageName: ""))
   
   private let contentsView: UIView = {
     let view = UIView()
@@ -42,7 +42,7 @@ final class CategoryView: UIScrollView {
   }
 }
 
-extension CategoryView: BaseView {
+extension FavorCategoryView: BaseView {
   func setupStyles() {
     self.backgroundColor = .favorColor(.background)
     self.showsHorizontalScrollIndicator = false
