+++ Favor/Favor/Sources/Scenes/Gift/ViewControllers/NewGiftVC.swift
@@ -146,14 +146,14 @@ final class NewGiftViewController: BaseViewController {
   private lazy var titleLabel = self.makeTitleLabel("제목")
   private lazy var titleLine = self.makeLine()
   private lazy var categoryLabel = self.makeTitleLabel("카테고리")
-  private let categoryView = CategoryView()
+  private let categoryView = FavorCategoryView()
   private lazy var pictureLabel = self.makeTitleLabel("사진")
   private lazy var friendLabel = self.makeTitleLabel("준 사람")
-  private let giverButton = PlainFavorButton(with: .main("친구 선택", isRight: true))
-  private let receiverButton = PlainFavorButton(with: .main("친구선택", isRight: true))
+  private let giverButton = FavorPlainButton(with: .main("친구 선택", isRight: true))
+  private let receiverButton = FavorPlainButton(with: .main("친구선택", isRight: true))
   private lazy var friendLine = self.makeLine()
   private lazy var dateLabel = self.makeTitleLabel("날짜")
-  private let dateButton = PlainFavorButton(with: .main("날짜 선택", isRight: false))
+  private let dateButton = FavorPlainButton(with: .main("날짜 선택", isRight: false))
   private lazy var dateLine = self.makeLine()
   private lazy var emotionLabel = self.makeTitleLabel("감정 메모")
   private lazy var emotionButton1 = self.makeEmotionButton("🥹")
